final_cut_composer.py
# final_cut_composer.py
from typing import List, Dict, Any, Optional
import google.generativeai as genai # 만약 이 모듈도 LLM을 직접 사용한다면 필요

class FinalCutComposer:
    def __init__(self, gemini_model: Optional[genai.GenerativeModel] = None):
        self.model = gemini_model # LLM을 활용한 말풍선 스타일 제안, 최종 검토 등에 사용 가능
        if self.model:
            print("FinalCutComposer initialized with Gemini Model.")
        else:
            print("FinalCutComposer initialized W/O Gemini Model.")

    def add_dialogues_to_image(self,
                               base_image_data: bytes,
                               dialogues_for_cut: List[Dict[str, Any]],
                               speech_bubble_guidance_for_cut: List[Dict[str, Any]],
                               font_and_bubble_styles: Optional[Dict[str, Any]] = None) -> bytes:
        print(f"\n[Composer] Adding {len(dialogues_for_cut)} dialogues to image (mocked data length: {len(base_image_data)} bytes).")

        if not dialogues_for_cut:
            print("  -> No dialogues to add. Returning base image.")
            return base_image_data

        # 여기에 실제 이미지 처리 라이브러리(Pillow, OpenCV 등)를 사용한 로직 구현
        # 1. base_image_data를 이미지 객체로 로드
        # 2. dialogues_for_cut, speech_bubble_guidance_for_cut 정보를 바탕으로 말풍선과 텍스트 렌더링
        #    - self.model(LLM)을 사용하여 대사의 감정/뉘앙스에 따른 말풍선 스타일을 구체화할 수 있음
        #    - 이미지 분석을 통해 캐릭터 위치를 파악하고 말풍선 위치를 최종 조정할 수 있음
        # 3. 수정된 이미지 객체를 바이트 데이터로 변환하여 반환
        
        final_text_overlay = ["\n--- Applied Dialogues ---"]
        for i, dialogue in enumerate(dialogues_for_cut):
            guidance = next((g for g in speech_bubble_guidance_for_cut if g.get('dialogue_id') == dialogue.get('id')), {})
            text_to_add = (
                f"  Dialogue ID: {dialogue.get('id', 'N/A')}, "
                f"Speaker: {dialogue.get('speaker', 'N/A')}, "
                f"Text: \"{dialogue.get('text', 'N/A')}\" (Nuance: {dialogue.get('nuance', 'N/A')}), "
                f"Bubble Suggestion: {guidance.get('suggested_area', 'N/A')} "
                f"({guidance.get('bubble_style_hint', 'N/A')})"
            )
            final_text_overlay.append(text_to_add)
        
        # 원본 이미지 바이트와 텍스트 설명을 합쳐서 반환 (실제로는 이미지에 그려야 함)
        final_webtoon_cut_image_data = base_image_data + "\n".join(final_text_overlay).encode('utf-8')
        
        print(f"  -> Dialogues added (conceptually). Final data length: {len(final_webtoon_cut_image_data)} bytes.")
        return final_webtoon_cut_image_data