requirement.txt
google-generativeai
google-cloud-aiplatform
requests